+++ app/assets/frontend/components/Tweet.jsx
@@ -1,3 +1,4 @@
+import React from 'react';
 export default class Tweet extends React.Component{
   render(){
     return(
