+++ app/assets/frontend/components/TweetBox.jsx
@@ -1,4 +1,6 @@
+import React from 'react';
 import TweetActions from "../actions/TweetActions";
+
 
 export default class TweetBox extends React.Component{
   sendTweet(event){
