+++ app/assets/frontend/components/TweetList.jsx
@@ -1,4 +1,5 @@
-import Tweet from "./Tweet"
+import React from 'react';
+import Tweet from "./Tweet";
 
 export default class TweetList extends React.Component{
   render(){
