+++ app/assets/javascripts/application.js
@@ -12,6 +12,4 @@
 //
 //= require jquery
 //= require jquery_ujs
-//= require react
-//= require react_ujs
 //= require_tree .
